Move socket logic and state common for all games to a seperate hook useGame from Soccer game
